// src/user/dtoss/update.dto.ts
export class UpdateUserDto { 
    firstname: string; 
    email: string;
    password: string;
    lastname: string;
    uploadresume: string;
    conntactmethod: string;
    contacttime: string;
    home: string;
    work: string;
    mobile: string;
    employmenttype: string;
    salarexpection: string;
    whencanyoustart: string;
    eduaction: string;
    employment: string;
    educationdegree: string;
    specialization: string;
    gradeGpa: string;
    compleatedate: string;
    Contactname: string;
    country: string;
    city: string;
    zipcode: string;
}